Header added

+++ tests/Utils.h
@@ -1,4 +1,5 @@
 
+#include <algorithm>
 #include <numeric>
 #include <vector>
 
@@ -7,11 +8,11 @@
 
 static void checkTiles(uint size, const std::vector<size_t>& tiles)
 {
-    uint tilesCount = size * size;
-    std::vector<uint> testValues(tilesCount);
+    size_t tilesCount = size * size;
+    std::vector<size_t> testValues(tilesCount);
     std::iota(testValues.begin(), testValues.end(), 0);
 
-    for (uint i = 0; i < tilesCount; i++)
+    for (size_t i = 0; i < tilesCount; i++)
     {
         EXPECT_NE(std::find(testValues.begin(), testValues.end(), tiles[i]), testValues.end());
     }
